Adds RAILS_SERVE_STATIC_FILES environment variable during launch
Without this, Rails expects a static file server like NGINX or HTTPD to
serve the static assets that are generated by this buildpack. The
buildpack should allow that, but also not assume that there will be a
static fileserver in front of the rails process. By settings the
RAILS_SERVE_STATIC_FILES environment variable, the Rails process will
handle requests for these static files.